student home에서 더미 데이터를 실제 데이터로 변경합니다
